backend/requirements.txt
Django>=6.0.1
djangorestframework>=3.15.1
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
drf-yasg>=1.21.7
python-dotenv>=1.0.1
Pillow>=10.3.0
gunicorn>=21.2.0
whitenoise>=6.6.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
